reformat: file hierarchy
