+++ .gitignore
@@ -2,7 +2,7 @@
 __pycache__/
 *.py[cod]
 *$py.class
-/.idea
+#/.idea
 # C extensions
 *.so
 
@@ -157,5 +157,9 @@ cython_debug/
 #  be found at https://github.com/github/gitignore/blob/main/Global/JetBrains.gitignore
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
-#.idea/
-Module 1/Week 1/.idea/discord.xml
+.idea/
+#Module 1/Week 1/.idea/discord.xml
+#Module 1/Week 1/.idea/vcs.xml
+#Module 1/Week 1/.idea/misc.xml
+#Module 1/Week 1/.idea/inspectionProfiles/Project_Default.xml
+#Module 1/Week 1/.idea/inspectionProfiles/profiles_settings.xml
